+++ src/main/java/com/pive/empickbackend/employment/jobtest/command/application/controller/JobtestCommandController.java
@@ -1,0 +1,4 @@
+package com.pive.empickbackend.employment.jobtest.command.application.controller;
+
+public class JobtestCommandController {
+}
