+++ src/main/java/com/pive/empickbackend/employment/jobtest/command/application/dto/JobtestCommandDTO.java
@@ -1,0 +1,4 @@
+package com.pive.empickbackend.employment.jobtest.command.application.dto;
+
+public class JobtestCommandDTO {
+}
