+++ src/main/java/com/pive/empickbackend/employment/jobtest/command/application/service/JobtestCommandService.java
@@ -1,0 +1,4 @@
+package com.pive.empickbackend.employment.jobtest.command.application.service;
+
+public interface JobtestCommandService {
+}
