+++ src/main/java/com/pive/empickbackend/employment/jobtest/command/application/service/JobtestCommandServiceImpl.java
@@ -1,0 +1,4 @@
+package com.pive.empickbackend.employment.jobtest.command.application.service;
+
+public class JobtestCommandServiceImpl {
+}
