+++ src/main/java/com/pive/empickbackend/employment/jobtest/command/domain/aggregate/JobtestEntity.java
@@ -1,0 +1,4 @@
+package com.pive.empickbackend.employment.jobtest.command.domain.aggregate;
+
+public class JobtestEntity {
+}
