+++ src/main/java/com/pive/empickbackend/employment/jobtest/command/domain/repository/JobtestRepository.java
@@ -1,0 +1,4 @@
+package com.pive.empickbackend.employment.jobtest.command.domain.repository;
+
+public class JobtestRepository {
+}
