+++ src/main/java/com/pive/empickbackend/employment/jobtest/query/controller/JobtestQueryController.java
@@ -1,0 +1,4 @@
+package com.pive.empickbackend.employment.jobtest.query.controller;
+
+public class JobtestQueryController {
+}
