+++ src/main/java/com/pive/empickbackend/employment/jobtest/query/dto/JobtestQueryDTO.java
@@ -1,0 +1,4 @@
+package com.pive.empickbackend.employment.jobtest.query.dto;
+
+public class JobtestQueryDTO {
+}
