+++ src/main/java/com/pive/empickbackend/employment/jobtest/query/mapper/JobtestMapper.java
@@ -1,0 +1,4 @@
+package com.pive.empickbackend.employment.jobtest.query.mapper;
+
+public class JobtestMapper {
+}
