+++ src/main/java/com/pive/empickbackend/employment/jobtest/query/service/JobtestQueryServiceImpl.java
@@ -1,0 +1,4 @@
+package com.pive.empickbackend.employment.jobtest.query.service;
+
+public class JobtestQueryServiceImpl {
+}
