--- src/main/java/com/piveguyz/empickbackend/employment/jobtest/command/application/controller/JobtestCommandController.java
@@ -1,4 +1,0 @@
-package com.piveguyz.empickbackend.employment.jobtest.command.application.controller;
-
-public class JobtestCommandController {
-}
