--- src/main/java/com/piveguyz/empickbackend/employment/jobtest/command/application/dto/JobtestCommandDTO.java
@@ -1,4 +1,0 @@
-package com.piveguyz.empickbackend.employment.jobtest.command.application.dto;
-
-public class JobtestCommandDTO {
-}
