--- src/main/java/com/piveguyz/empickbackend/employment/jobtest/command/application/service/JobtestCommandServiceImpl.java
@@ -1,4 +1,0 @@
-package com.piveguyz.empickbackend.employment.jobtest.command.application.service;
-
-public class JobtestCommandServiceImpl {
-}
