--- src/main/java/com/piveguyz/empickbackend/employment/jobtest/command/domain/aggregate/JobtestEntity.java
@@ -1,4 +1,0 @@
-package com.piveguyz.empickbackend.employment.jobtest.command.domain.aggregate;
-
-public class JobtestEntity {
-}
