--- src/main/java/com/piveguyz/empickbackend/employment/jobtest/command/domain/repository/JobtestRepository.java
@@ -1,4 +1,0 @@
-package com.piveguyz.empickbackend.employment.jobtest.command.domain.repository;
-
-public class JobtestRepository {
-}
