--- src/main/java/com/piveguyz/empickbackend/employment/jobtest/query/controller/JobtestQueryController.java
@@ -1,4 +1,0 @@
-package com.piveguyz.empickbackend.employment.jobtest.query.controller;
-
-public class JobtestQueryController {
-}
