--- src/main/java/com/piveguyz/empickbackend/employment/jobtest/query/dto/JobtestQueryDTO.java
@@ -1,4 +1,0 @@
-package com.piveguyz.empickbackend.employment.jobtest.query.dto;
-
-public class JobtestQueryDTO {
-}
