--- src/main/java/com/piveguyz/empickbackend/employment/jobtest/query/mapper/JobtestMapper.java
@@ -1,4 +1,0 @@
-package com.piveguyz.empickbackend.employment.jobtest.query.mapper;
-
-public class JobtestMapper {
-}
