--- src/main/java/com/piveguyz/empickbackend/employment/jobtest/query/service/JobtestQueryServiceImpl.java
@@ -1,4 +1,0 @@
-package com.piveguyz.empickbackend.employment.jobtest.query.service;
-
-public class JobtestQueryServiceImpl {
-}
